Re-separate dice app, revamp landing page
