+++ dice/static/dice/dice.js
