+++ dice/tests.py
