+++ dice/urls.py
@@ -2,6 +2,7 @@ from django.conf.urls import url
 
 from . import views
 
+app_name = 'dice'
 urlpatterns = [
-    url(r'^$', views.index, name='index'),
+    url(r'^$', views.index, name='index')
 ]
