+++ dice/views.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+from django.http import HttpResponse
+from django.template import loader
+# Create your views here.
+def index(request):
+	return render(request, 'dice/index.html')
