--- front/views.py
@@ -1,8 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-from django.http import HttpResponse
-
-
-def index(request):
-    return HttpResponse("Gloriously DECIDUOUSLY")
