+++ landing/static/landing/style.css
@@ -1,0 +1,18 @@
+body {
+	background-image: 	url('frost.jpg');
+	background-repeat: 	no-repeat;
+	background-size: 	100%;
+	color: 				brown;
+}
+
+#header {
+	position: 	absolute;
+	text-align: center;
+	bottom: 	70px;
+	right: 		550px;
+	width: 		200px;
+}
+
+a:visited {
+	color: orange-red;
+}
