+++ landing/views.py
@@ -2,14 +2,13 @@ from django.shortcuts import get_object_or_404, render
 from django.http import HttpResponse
 from django.template import loader
 
-from .models import App
+from .models import App, Post
 
 # Create your views here.
 def index(request):
+	posts = Post.objects.all()
+	context = {'posts': posts}
 	return render(request, 'landing/index.html')
-
-def dice(request):
-	return render(request, 'landing/dice.html')
 
 def list(request):
 	apps = App.objects.all()
